src/django/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
apscheduler==3.6.3
    # via python-telegram-bot
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
backports-zoneinfo==0.2.1
    # via
    #   django
    #   django-celery-beat
    #   django-timezone-field
    #   tzlocal
billiard==3.6.4.0
    # via celery
cachetools==4.2.2
    # via
    #   google-auth
    #   python-telegram-bot
celery==5.2.7
    # via
    #   django-celery-beat
    #   django-celery-results
certifi==2022.6.15
    # via
    #   python-telegram-bot
    #   requests
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
deprecated==1.2.13
    # via redis
django==4.0.6
    # via
    #   -r requirements.in
    #   django-bulk-update-or-create
    #   django-celery-beat
    #   django-cors-headers
    #   djangorestframework
django-bulk-update-or-create==0.3.0
    # via -r requirements.in
django-celery-beat==2.3.0
    # via -r requirements.in
django-celery-results==2.4.0
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-timezone-field==5.0
    # via django-celery-beat
djangorestframework==3.13.1
    # via -r requirements.in
google-auth==2.9.1
    # via
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==0.5.2
    # via gspread
gspread==5.4.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.20.4
    # via oauth2client
idna==3.3
    # via requests
kombu==5.2.4
    # via celery
oauth2client==4.1.3
    # via -r requirements.in
oauthlib==3.2.0
    # via requests-oauthlib
packaging==21.3
    # via redis
prettytable==3.3.0
    # via -r requirements.in
prompt-toolkit==3.0.30
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   oauth2client
pyparsing==3.0.9
    # via
    #   httplib2
    #   packaging
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
python-dotenv==0.20.0
    # via -r requirements.in
python-telegram-bot==13.13
    # via -r requirements.in
pytz==2022.1
    # via
    #   apscheduler
    #   celery
    #   django-timezone-field
    #   djangorestframework
    #   python-telegram-bot
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via -r requirements.in
redis==4.3.4
    # via -r requirements.in
requests==2.28.1
    # via
    #   -r requirements.in
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.8
    # via
    #   google-auth
    #   oauth2client
six==1.16.0
    # via
    #   apscheduler
    #   click-repl
    #   google-auth
    #   oauth2client
    #   python-dateutil
sqlparse==0.4.2
    # via django
tornado==6.2
    # via python-telegram-bot
tzdata==2022.1
    # via django-celery-beat
tzlocal==4.2
    # via apscheduler
urllib3==1.26.10
    # via requests
vine==5.0.0
    # via
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prettytable
wrapt==1.14.1
    # via deprecated
xmltodict==0.13.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
